refactor(cuda): reorganize CUDA engines
- Splits files and functions related to the low latency and amortized
engines.
